+++ src/helpers/DynamicMeta.php
@@ -189,10 +189,12 @@ class DynamicMeta
                     if (is_array($href)) {
                         $headerValue = '';
                         foreach ($href as $hrefVal) {
+                            $hrefVal = UrlHelper::encodeUrl($hrefVal);
                             $headerValue .= ('<' . $hrefVal . '>' . ',');
                         }
                         $headerValue = rtrim($headerValue, ',');
                     } else {
+                        $href = UrlHelper::encodeUrl($href);
                         $headerValue = '<' . $href . '>';
                     }
                     $headerValue .= "; rel='canonical'";